Create README.md
Synopsis
The management team at AAVAIL is preparing to deploy a large number of teams each tasked with integration into a different new market. They claim to have a optimized the teams fairly with respect to skills and experience. They are asking you to come up with a framework to evaluate the makeup of their teams. They have not finished hiring and creating all of the teams so naturally before you even get the data you wanted to get a head start.

Getting a head start usually involves finding a similar dataset and writing the code in a way that the new data, once obtained can be added with little effort.

When we perform a large number of statistical tests, some will have 𝑝
-values less than the designated level of 𝛼
 (e.g. 0.05) purely by chance, even if all the null hypotheses are really true. This is an inherent risk of using inferrential statistics. Fortunately, there are several techniques to mitigate the risk.

We are going to look at the 2018 world cup data in this example.

The case study is comprised of the following sections:

Data Cleaning
Data Visualization
NHT
Adjust NHT results for multiple comparisons
Data science work that focuses on creating a predictive model is perhaps the hallmark of the field today, but there are still many use cases where inferential statistics are the best tool available. One issue with statistical inference is that there are situations where performing multiple tests is a logical way to accomplish a task, but it comes at the expense of an increased rate of false positives or Type I errors.

In this case study you will apply techniques and knowledge from all of the units in Module 2.

Getting started
This unit is interactive. During this unit we encourage you to open this file as a notebook. Download the notebook from the following link then open it locally using a Jupyter server or use your IBM cloud account to login to Watson Studio. Inside of Waston Studio cloud if you have not already ensure that this notebook is loaded as part of the project for this course. As a reminder fill in all of the places in this notebook marked with YOUR CODE HERE or YOUR ANSWER HERE. The data and notebook for this unit are available below.

m2-u7-case-study.ipynb
worldcup-2018.csv
This unit is organized into the following sections:

Data Processing
Data Summary
Investigative Visualization
Hypothesis testing
Resources
Creating or uploading a notebook in IBM cloud
Resources for multiple testing in Python